requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.42.2
colorama==0.4.6
comm==0.2.2
comtypes==1.4.6
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
flask-cors==5.0.1
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.8
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.5.0
kiwisolver==1.4.5
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mlxtend==0.23.4
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.41.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.78.1
opencv-python==4.11.0.86
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.1
protobuf==4.25.5
psutil==6.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyjokes==0.8.3
pymongo==4.13.2
pyparsing==3.1.4
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
pyttsx3==2.91
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.33.0
Send2Trash==1.8.3
setuptools==74.0.0
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
streamlit==1.45.1
sympy==1.13.2
tenacity==9.1.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.3.0
tokenizers==0.21.0
toml==0.10.2
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.49.0
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20240821
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0
